+++ src/main/java/net/spell_engine/api/render/StunParticleSpawner.java
@@ -2,11 +2,28 @@ package net.spell_engine.api.render;
 
 import net.minecraft.client.world.ClientWorld;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.particle.ParticleEffect;
+import net.minecraft.particle.ParticleType;
 import net.minecraft.particle.ParticleTypes;
+import net.minecraft.registry.Registries;
+import net.minecraft.util.Identifier;
 import net.minecraft.util.math.Vec3d;
 import net.spell_engine.api.effect.CustomParticleStatusEffect;
 
 public class StunParticleSpawner implements CustomParticleStatusEffect.Spawner {
+
+    private final ParticleEffect particleEffect;
+    public StunParticleSpawner() {
+        this.particleEffect = ParticleTypes.CRIT;
+    }
+
+    public StunParticleSpawner(Identifier particleId) {
+        var particleType = (ParticleEffect) Registries.PARTICLE_TYPE.get(particleId);
+        if (particleType == null) {
+            throw new IllegalArgumentException("Particle type not found: " + particleId);
+        }
+        this.particleEffect = particleType;
+    }
 
     @Override
     public void spawnParticles(LivingEntity livingEntity, int amplifier) {
@@ -20,7 +37,7 @@ public class StunParticleSpawner implements CustomParticleStatusEffect.Spawner {
                     .add(0, livingEntity.getHeight() * 1.2F, 0)
                     .add(rotated);
             // System.out.println("Spawning stun particle at angle: " + (angle % 360) + " time: " + time);
-            clientWorld.addParticle(ParticleTypes.CRIT, true, spawnPosition.x, spawnPosition.y, spawnPosition.z, 0, 0, 0);
+            clientWorld.addParticle(particleEffect, true, spawnPosition.x, spawnPosition.y, spawnPosition.z, 0, 0, 0);
         }
     }
 }